Change colors from integers to floats
The color used to be stored as integers in the range [0..255], but now
are stored as floats in the range [0.0..100.0]. This is so that we can
express the color as percentages.

+++ src/color.rs
@@ -1,15 +1,15 @@
 pub struct Color {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
 }
 
 impl Color {
     pub fn new(rs: String, gs: String, bs: String) -> Color {
         Color {
-            r: rs.parse::<u8>().unwrap(),
-            g: gs.parse::<u8>().unwrap(),
-            b: bs.parse::<u8>().unwrap(),
+            r: rs.parse::<f32>().unwrap(),
+            g: gs.parse::<f32>().unwrap(),
+            b: bs.parse::<f32>().unwrap(),
         }
     }
 }
